split out inventory into its own file
